added map componenet
