+++ covid-resource-guide/package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "dotenv": "^10.0.0",
     "mapbox-gl": "^2.6.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
